Change YoinkPaste_p and YoinkPaste_P mappings
The previous commands did not work properly in some cases, like if you
are in Insert mode, and then press <c-o> to enter Normal mode temporarily,
then only the yoink#setupPaste part actually ran and the remaining of the command got
entered in the buffer (since that mode allows us to enter ONE command only)
I had also seen this behaviour in macro-recording mode.

The new mappings run the commands all in one go using the execute()
function, so the previous artifacts are now avoided.